\name{DmIALite}
\alias{DmIALite}
\docType{data}
\title{ Fly gene interaction matrix }
\description{
  Preprocessed fly gene interaction matrix
}
\usage{data(DmIALite)}
\details{
  a 4-column matrix containing fly interacted genes and evidences
}
\references{ Feng, G., Du, P., Krett, N., Tessel, M., Rosen, S., Kibbe, W.A. and Lin, S.M., 'A collection of bioconductor methods to visualize gene-list annotations', BMC Research Notes 2010, 3:10}
\examples{
data(DmIALite)
DmIALite[1:4,]
}
\keyword{datasets}
