\name{DOLite}
\alias{DOLite}
\docType{data}
\title{ Disease Ontology Annotation List}
\description{
  Disease Ontology Annotation List
}
\usage{data(DOLite)}
\details{
  a standard list, whose names are DOLite IDs and each element contains the gene Entrez IDs belonging to the corresponding DOLite IDs.
}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  Du, P., Feng, G., Flatow, J., Song, J., Holko, M., Kibbe, W.A. and Lin, S.M., (2009) 'From disease ontology to disease-ontology lite: statistical methods to adapt a general-purpose ontology for the test of gene-ontology associations', Bioinformatics 25(12):i63-8
}
\examples{
data(DOLite)
DOLite[1:2]
}
\keyword{datasets}
