\name{DOLiteTerm}
\alias{DOLiteTerm}
\docType{data}
\title{ Disease Ontology Annotation Vector}
\description{
  Disease Ontology Annotation Vector
}
\usage{data(DOLiteTerm)}
\format{
  
}
\details{
  a character vector, where names are DOLite IDs and elements are Terms
}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  Du, P., Feng, G., Flatow, J., Song, J., Holko, M., Kibbe, W.A. and Lin, S.M., (2009) 'From disease ontology to disease-ontology lite: statistical methods to adapt a general-purpose ontology for the test of gene-ontology associations', Bioinformatics 25(12):i63-8
}
\examples{
data(DOLiteTerm)
DOLiteTerm[1:10]
}
\keyword{datasets}
