\name{drawTable}
\alias{drawTable}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Concept-Gene Networking Plotting }
\description{
  A function to generate a multigroup concepts-genes table 
}
\usage{drawTable(dataMatrix, topCat=10, heatMap=TRUE, matrixOfHeatmap=NULL, clusterTable=c('geneNum', 'pvalue', NULL), methodOfCluster=c('mds', 'sort'), mar=c(1,5,5,8), 
			addRowLabel=TRUE, cex.axis=c(1.1, 0.9), reverseOfCluster=FALSE, xGridLine=FALSE, colorBar=TRUE, newWindow=TRUE, endOfColoBar=c('> 0.01', 'Minimum of p values'), ...)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataMatrix}{ a top concepts-genes matrix generated by \code{\link{getConceptTable}}. }
  \item{topCat}{ number to specify how many top concepts-genes analysis will show. }
  \item{heatMap}{ logic, determine whether the multiple group concepts-genes table is presented by heatmap. }
  \item{matrixOfHeatmap}{ NULL or a concepts-genes matrix generated by \code{\link{getConceptTable}}, which is used to show enrichment test significance for each concept. }
  \item{clusterTable}{ cluster data to specify which type of values will be used for cluster. }
  \item{methodOfCluster}{ cluster method}
  \item{mar}{ marginal parameter for table, please see \code{\link{par}}}
  \item{addRowLabel}{ logic, whether add row names}
  \item{cex.axis}{ font size parameter for table, please see \code{\link{par}} }
  \item{reverseOfCluster}{ logic, whether reverse the cluster order. }
  \item{xGridLine}{ logic, whether add horizontal line in table or not}
  \item{colorBar}{ logic, whether show color bar or not}
  \item{newWindow}{ logic, whether present table in current active window or not} 
  \item{endOfColoBar}{ a character string for color bar.} 
  \item{\dots}{ other parameters used by 'sort' }
}
\details{
  an image based multigroup concepts-genes table is generated. If heatmap is on, the statistical significant cells are shaded by different level green. Specified top gene amounts are highlighted as red.   
}
\value{
  No return value. 
}
\references{ Feng, G., Du, P., Krett, N., Tessel, M., Rosen, S., Kibbe, W.A. and Lin, S.M., 'A collection of bioconductor methods to visualize gene-list annotations', BMC Research Notes 2010, 3:10}
\author{ Gang Feng, Pan Du and Simon Lin }
\seealso{ See Also as \code{\link{getConceptTable}}, \code{\link{groupReport}} }
\examples{
data(sampleGroupsData)
gAKEGGL <- lapply(sampleGroupsData, geneAnswersBuilder, 'org.Hs.eg.db', categoryType='KEGG', pvalueT=0.1, verbose=FALSE)
#output<- getConceptTable(gAKEGGL, items='geneNum')
\dontrun{drawTable(output[[1]], matrixOfHeatmap=output[[2]], mar=c(2,15,3,2), clusterTable=NULL)} 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
