\name{getListGIF}
\alias{getListGIF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{getListGIF}
\description{
  Function to grab the overrepresented concepts from a list of genes throught ListGIF web server.
}
\usage{
getListGIF(glist, output=NULL, background=c("white", "black"), type=c("genelevel", "riflevel", "goterm"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{glist}{a character vector or a list or an one-column data frame giving gene symbols for query}
  \item{output}{a character string specifying the name of the word cloud figure. Default is NULL, in this case the file name is generated by ListGIF web server}
  \item{background}{a character string specifying the background of the word cloud png file to be either "white"" or "black". Default is "white""}
  \item{type}{a character string specifying the type of analysis implemented by ListGIF web server: "genelevel", "riflevel" or "goterm". Default is "genelevel" }
}
\details{
  glist is a character vector of gene symbols in query. If glist is a list or a one-column data frame, it will be converted to a vector. 
  output is the file name of the word cloud figure. If not specified, the original file name generated by ListGIF server is provided.
  background of the word cloud figure can be either white or black. Default is white.
  type specifies the word cloud to be eigher geneGIF, geneRIF or GO terms. geneRIF analysis requires considerable amount of time.
}
\value{
A png file is generated at current working directory if user-specified directory is not provied in output argument.
}
\references{
Wen, J., Wang X., Kibbe, W., Lin S., Lu, H., 'Visual Annotation of the Gene Database', Conf Proc IEEE Eng Med Biol Soc. 2009;2009:4175-7
}
\author{
Lei Huang
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{ \code{\link{help}}}
\examples{
glist <- c("MRPS35","NBL1","PSMD14","PGK1","SMC4","SLC16A1","CAV1")
\dontrun{getListGIF(glist=glist)}
\dontrun{getListGIF(glist=glist, output="tmp2.png", background="black", type="riflevel")}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

