\name{groupReport}
\alias{groupReport}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Generate a multigroup Concepts-genes analysis report }
\description{
  Function to generate a html format top multigroup Concepts-genes analysis report based on a given GeneAnswers instance list. Currently, this function only works at Mac platform and Java 1.6 or latest version is required.
}
\usage{
groupReport(dataMatrix, gAList, topCat=10, methodOfCluster=c('mds', 'sort'), matrixOfHeatmap=NULL, clusterTable=c('geneNum', 'pvalue', NULL), catTerm=TRUE,
		fileName = "multiConceptsGenes.html", title='Multigroup Genes Concepts Analysis', catType=c('GO', 'KEGG', 'DOLITE', 'REACTOME.PATH', 'CABIO.PATH', 'Unknown'),
		reverseOfCluster=FALSE,  colorValueColumn = NULL, annLib=c('org.Hs.eg.db', 'org.Rn.eg.db', 'org.Mm.eg.db', 'org.Dm.eg.db'), nameLength=94, addID=TRUE, interactive=FALSE,
		bgColor='#ffffcc', keepCytoscapeFiles=TRUE, ...)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataMatrix}{ a top concepts-genes matrix generated by \code{\link{getConceptTable}}. }
  \item{gAList}{ a GeneAnswers instance list. } 
  \item{topCat}{ number to specify how many top concepts-genes analysis will show. }
  \item{methodOfCluster}{ cluster method}
  \item{matrixOfHeatmap}{ NULL or a concepts-genes matrix generated by \code{\link{getConceptTable}}, which is used to show enrichment test significance for each concept. }
  \item{clusterTable}{ cluster data to specify which type of values will be used for cluster. }
  \item{catTerm}{ logic, determine whether mapping category IDs to names}
  \item{fileName}{ output html file name }
  \item{title}{ output html title }
  \item{catType}{ category type, current version supports 'GO', 'KEGG', 'DOLITE', 'REACTOME.PATH', 'CABIO.PATH' and customized annotation libraries, 'Unknown'. }
  \item{reverseOfCluster}{ logic, whether reverse the cluster order. }
  \item{colorValueColumn}{ numbers or column names of geneInput slots of the given GeneAnswers instance list to specify the colors of leaves }
  \item{annLib}{ annotation librarry names, current version supports 'org.Hs.eg.db', 'org.Rn.eg.db', 'org.Mm.eg.db' and 'org.Dm.eg.db'. }
  \item{nameLength}{ show how many first letters for long term names, 'all' for full name, default value is 94. }
  \item{addID}{ logic, add term IDs following term names or not}
  \item{interactive}{ logic, determine whether network is interactive or not. Interactive network requires java and flash supports. }
  \item{bgColor}{ a R compatible color for html background color.}
  \item{keepCytoscapeFiles}{ logic, determine whether to keep cytoscape files if interactive is set to TRUE}
  \item{\dots}{ other parameters used by 'sort' }
}
\details{
  In general, a html format top multigroup Concepts-genes analysis report is generated. It includes a multigroup concepts-genes table, several concepts-genes networks figures and a couple of tables
containing genes and their information.
  colorValueColumn could be NULL, column name or a same length column-name vector as length of the given GeneAnswers instantce list. No color for genes if it is NULL. All of GeneAnswers instances
are applied color for genes based on the same column name if the length is one. Or the colors of genes in concepts-genes networks are based on the same length column-name vector. 
  If catType is not set to 'Unknown', catTerm in function getConceptTable should be set to FALSE.
}
\value{
  no value returned
}
\references{ Feng, G., Du, P., Krett, N., Tessel, M., Rosen, S., Kibbe, W.A. and Lin, S.M., 'A collection of bioconductor methods to visualize gene-list annotations', BMC Research Notes 2010, 3:10}
\author{ Gang Feng, Pan Du and Simon Lin }
\seealso{\code{\link{getConceptTable}}, \code{\link{drawTable}}}
\examples{
data(sampleGroupsData)
gAKEGGL <- lapply(sampleGroupsData, geneAnswersBuilder, 'org.Hs.eg.db', categoryType='KEGG', pvalueT=0.1, verbose=FALSE)
output<- getConceptTable(gAKEGGL, catTerm=FALSE, items='geneNum')
groupReport(output[[1]], gAKEGGL,  matrixOfHeatmap=output[[2]], clusterTable=NULL, fileName='KEGGtest.html',  catType='KEGG', colorValueColumn=colnames(sampleGroupsData[[1]])[-1]) 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
