\name{HsIALite}
\alias{HsIALite}
\docType{data}
\title{ Human gene interaction matrix  }
\description{
  Preprocessed human gene interaction matrix
}
\usage{data(HsIALite)}
\details{
  a 4-column matrix containing human interacted genes and evidences
}
\references{ Feng, G., Du, P., Krett, N., Tessel, M., Rosen, S., Kibbe, W.A. and Lin, S.M., 'A collection of bioconductor methods to visualize gene-list annotations', BMC Research Notes 2010, 3:10}
\examples{
data(HsIALite)
HsIALite[1:4,]
}
\keyword{datasets}
