\name{humanExpr}
\alias{humanExpr}
\docType{data}
\title{ Example human expression data}
\description{
  An example data of human expression 
}
\usage{data(humanExpr)}
\format{
  A data frame with 86 observations on the 6 variables.
}
\details{
  This date frame is a part of expression profile from a human Illumina array experiment.
}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  ~~ possibly secondary sources and usages ~~
}
\examples{
data(humanExpr)
humanExpr[1:10,]
}
\keyword{datasets}
