\name{MmIALite}
\alias{MmIALite}
\docType{data}
\title{ Mouse gene interaction matrix }
\description{
  Preprocessed mouse gene interaction matrix
}
\usage{data(MmIALite)}
\details{
  a 4-column matrix containing mouse interacted genes and evidences
}
\references{ Feng, G., Du, P., Krett, N., Tessel, M., Rosen, S., Kibbe, W.A. and Lin, S.M., 'A collection of bioconductor methods to visualize gene-list annotations', BMC Research Notes 2010, 3:10}
\examples{
data(MmIALite)
MmIALite[1:4,]
}
\keyword{datasets}
