\name{mouseExpr}
\alias{mouseExpr}
\docType{data}
\title{ Example mouse expression data }
\description{
  Example data of mouse expression
}
\usage{data(mouseExpr)}
\format{
  A data frame with 71 observations on the following 6 variables.
}
\details{
  This date frame is a part of expression profile from a mouse Illumina array experiment.
}
\references{ Feng, G., Du, P., Krett, N., Tessel, M., Rosen, S., Kibbe, W.A. and Lin, S.M., 'A collection of bioconductor methods to visualize gene-list annotations', BMC Research Notes 2010, 3:10}
\examples{
data(mouseExpr)
mouseExpr[1:10,]
}
\keyword{datasets}
