\name{sampleGroupsData}
\alias{sampleGroupsData}
\docType{data}
\title{ Example human expression data}
\description{
  An example data of human expression 
}
\usage{data(sampleGroupsData)}
\format{
  A data frame list containing genes and fold changes from 6 different comparisons.
}
\details{
  This date frame is a part of expression profile from a group of human Illumina array experiments.
}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  ~~ possibly secondary sources and usages ~~
}
\examples{
data(sampleGroupsData)
head(sampleGroupsData)
}
\keyword{datasets}
