\name{humanExpr}
\alias{humanExpr}
\docType{data}
\title{ Example human expression data}
\description{
  An example data of human expression 
}
\usage{data(humanExpr)}
\format{
  A data frame with 86 observations on the 6 variables.
}
\details{
  This date frame is a part of expression profile from a human Illumina array experiment.
}
\references{ Feng, G., Du, P., Krett, N., Tessel, M., Rosen, S., Kibbe, W.A. and Lin, S.M., 'A collection of bioconductor methods to visualize gene-list annotations', BMC Research Notes 2010, 3:10}
\examples{
data(humanExpr)
humanExpr[1:10,]
}
\keyword{datasets}
