\name{mouseExpr}
\alias{mouseExpr}
\docType{data}
\title{ Example mouse expression data }
\description{
  Example data of mouse expression
}
\usage{data(mouseExpr)}
\format{
  A data frame with 71 observations on the following 6 variables.
}
\details{
  This date frame is a part of expression profile from a mouse Illumina array experiment.
}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  ~~ possibly secondary sources and usages ~~
}
\examples{
data(mouseExpr)
mouseExpr[1:10,]
}
\keyword{datasets}
