\name{sampleGroupsData}
\alias{sampleGroupsData}
\docType{data}
\title{ Example human expression data}
\description{
  An example data of human expression 
}
\usage{data(sampleGroupsData)}
\format{
  A data frame list containing genes and fold changes from 6 different comparisons.
}
\details{
  This date frame is a part of expression profile from a group of human Illumina array experiments.
}
\references{ Feng, G., Du, P., Krett, N., Tessel, M., Rosen, S., Kibbe, W.A. and Lin, S.M., 'A collection of bioconductor methods to visualize gene-list annotations', BMC Research Notes 2010, 3:10}
\examples{
data(sampleGroupsData)
head(sampleGroupsData)
}
\keyword{datasets}
